backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
Pillow==10.1.0
setuptools>=65.5.1
numpy==1.26.4
deepface==0.0.79
tensorflow==2.16.2
tf-keras>=2.15.0
groq>=0.11.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
